[#14] Usage of @ComponentDefinitionPostProcessor and @ComponentPostProcessor decorators - renamed @TimedDecorator to @Timed [Review by: @saskodh]
