+++ src/app/AppConfig.ts
@@ -4,7 +4,7 @@ import { RepositoriesConfig } from "./repositories/RepositoriesConfig";
 import { ServicesConfig } from "./services/ServicesConfig";
 import { InterceptorsConfig } from "./interceptors/InterceptorsConfig";
 import { ActiveProfiles } from "@sklechko/framework/lib/decorators/ProfileDecorators";
-import { PostProcessorsConfig } from "./postProcessors/PostProcessorsConfig";
+import { PostProcessorsConfig } from "./postProcessors/timed/PostProcessorsConfig";
 
 @Import(ControllersConfig, RepositoriesConfig, ServicesConfig, InterceptorsConfig, PostProcessorsConfig)
 @PropertySource(__dirname + '/../resources/app.properties.json')
